+++ actdiag/sphinxcontrib/actdiag.py
@@ -11,6 +11,7 @@
 """
 
 import posixpath
+import re
 import os
 import codecs
 try:
@@ -36,6 +37,18 @@ class ActdiagError(SphinxError):
 
 
 class Actdiag(ActdiagDirective):
+    def run(self):
+        try:
+            return super(Actdiag, self).run()
+        except diagparser.ParseException, e:
+            if self.content:
+                msg = '[%s] ParseError: %s\n%s' % (self.name, e, "\n".join(self.content))
+            else:
+                msg = '[%s] ParseError: %s\n%r' % (self.name, e, self.arguments[0])
+
+            reporter = self.state.document.reporter
+            return [reporter.warning(msg, line=self.lineno)]
+
     def node2image(self, node, diagram):
         return node
 
@@ -44,7 +57,7 @@ def get_image_filename(self, code, format, options, prefix='actdiag'):
     """
     Get path of output file.
     """
-    if format not in ('PNG', 'PDF'):
+    if format not in ('PNG', 'PDF', 'SVG'):
         raise ActdiagError('actdiag error:\nunknown format: %s\n' % format)
 
     if format == 'PDF':
@@ -108,6 +121,26 @@ def get_fontmap(self):
     return fontmap
 
 
+def get_anchor(self, refid, fromdocname):
+    for docname in self.builder.env.found_docs:
+        doctree = self.builder.env.get_doctree(docname)
+        for target in doctree.traverse(nodes.Targetable):
+            if target.attributes['refid'] == refid:
+                targetfile = self.builder.get_relative_uri(fromdocname, docname)
+                return targetfile + "#" + refid
+
+
+def resolve_reference(self, href, options):
+    if href is None:
+        return
+    pattern = re.compile(u"^:ref:`(.+?)`", re.UNICODE)
+    matched = pattern.search(href)
+    if matched:
+        return get_anchor(self, matched.group(1), options['current_docname'])
+    else:
+        return href
+
+
 def create_actdiag(self, code, format, filename, options, prefix='actdiag'):
     """
     Render actdiag code into a PNG output file.
@@ -118,22 +151,88 @@ def create_actdiag(self, code, format, filename, options, prefix='actdiag'):
         tree = diagparser.parse(diagparser.tokenize(code))
         screen = builder.ScreenNodeBuilder.build(tree)
 
+        for lane in screen.lanes:
+            if lane.href:
+                lane.href = resolve_reference(self, lane.href, options)
+        for node in screen.traverse_nodes():
+            if node.href:
+                node.href = resolve_reference(self, node.href, options)
+
         antialias = self.builder.config.actdiag_antialias
-        draw = DiagramDraw.DiagramDraw(format, screen, filename, fontmap=fontmap,
-                                       antialias=antialias)
+        draw = DiagramDraw.DiagramDraw(format, screen, filename,
+                                       fontmap=fontmap, antialias=antialias)
     except Exception, e:
         raise ActdiagError('actdiag error:\n%s\n' % e)
 
     return draw
 
 
+def make_svgtag(self, image, relfn, trelfn, outfn,
+                alt, thumb_size, image_size):
+    svgtag_format = """<svg xmlns="http://www.w3.org/2000/svg"
+    xmlns:xlink="http://www.w3.org/1999/xlink"
+    alt="%s" width="%s" height="%s">%s
+    </svg>"""
+
+    code = open(outfn, 'r').read().decode('utf-8')
+
+    return (svgtag_format %
+            (alt, image_size[0], image_size[1], code))
+
+
+def make_imgtag(self, image, relfn, trelfn, outfn,
+                alt, thumb_size, image_size):
+    result = ""
+
+    clickable_map = []
+    for l in image.diagram.lanes:
+        if l.href:
+            cell = image.metrics.lane_headerbox(l)
+            clickable_map.append((cell, l.href))
+    for n in image.nodes:
+        if n.href:
+            cell = image.metrics.cell(n)
+            clickable_map.append((cell, n.href))
+
+    if clickable_map:
+        imgtag_format = '<img src="%s" alt="%s" width="%s" '
+        imgtag_format += 'usemap="#map_1" height="%s" />\n'  # TODO:mapname
+    else:
+        imgtag_format = '<img src="%s" alt="%s" width="%s" height="%s" />\n'
+
+    if trelfn:
+        result += ('<a href="%s">' % relfn)
+        result += (imgtag_format %
+                   (trelfn, alt, thumb_size[0], thumb_size[1]))
+        result += ('</a>')
+    else:
+        result += (imgtag_format %
+                   (relfn, alt, image_size[0], image_size[1]))
+
+    if clickable_map:
+        result += ('<map name="map_1">')
+        rect_format = '<area shape="rect" coords="%s,%s,%s,%s" href="%s">'
+        for m in clickable_map:
+            x1 = m[0][0]
+            y1 = m[0][1]
+            x2 = m[0][2]
+            y2 = m[0][3]
+            result += (rect_format % (x1, y1, x2, y2, m[1]))
+
+        result += ('</map>')
+
+    return result
+
+
 def render_dot_html(self, node, code, options, prefix='actdiag',
                     imgcls=None, alt=None):
-    has_thumbnail = False
+    trelfn = None
+    thumb_size = None
     try:
         format = self.builder.config.actdiag_html_image_format
         relfn, outfn = get_image_filename(self, code, format, options, prefix)
 
+        options['current_docname'] = self.builder.current_docname
         image = create_actdiag(self, code, format, outfn, options, prefix)
         if not os.path.isfile(outfn):
             image.draw()
@@ -142,7 +241,6 @@ def render_dot_html(self, node, code, options, prefix='actdiag',
         # generate thumbnails
         image_size = image.pagesize()
         if 'maxwidth' in options and options['maxwidth'] < image_size[0]:
-            has_thumbnail = True
             thumb_prefix = prefix + '_thumb'
             trelfn, toutfn = get_image_filename(self, code, format,
                                                 options, thumb_prefix)
@@ -170,15 +268,13 @@ def render_dot_html(self, node, code, options, prefix='actdiag',
         if alt is None:
             alt = node.get('alt', self.encode(code).strip())
 
-        imgtag_format = '<img src="%s" alt="%s" width="%s" height="%s" />\n'
-        if has_thumbnail:
-            self.body.append('<a href="%s">' % relfn)
-            self.body.append(imgtag_format %
-                             (trelfn, alt, thumb_size[0], thumb_size[1]))
-            self.body.append('</a>')
+        if format == 'SVG':
+            tagfunc = make_svgtag
         else:
-            self.body.append(imgtag_format %
-                             (relfn, alt, image_size[0], image_size[1]))
+            tagfunc = make_imgtag
+
+        self.body.append(tagfunc(self, image, relfn, trelfn, outfn, alt,
+                                 thumb_size, image_size))
 
     self.body.append('</p>\n')
     raise nodes.SkipNode
@@ -215,7 +311,7 @@ def on_doctree_resolved(self, doctree, docname):
     if self.builder.name in ('gettext', 'singlehtml', 'html', 'latex'):
         return
 
-    for node in doctree.traverse(actdiag):  
+    for node in doctree.traverse(actdiag):
         code = node['code']
         prefix = 'actdiag'
         format = 'PNG'