+++ seqdiag/sphinxcontrib/seqdiag.py
@@ -36,6 +36,18 @@ class SeqdiagError(SphinxError):
 
 
 class Seqdiag(SeqdiagDirective):
+    def run(self):
+        try:
+            return super(Seqdiag, self).run()
+        except diagparser.ParseException, e:
+            if self.content:
+                msg = '[%s] ParseError: %s\n%s' % (self.name, e, "\n".join(self.content))
+            else:
+                msg = '[%s] ParseError: %s\n%r' % (self.name, e, self.arguments[0])
+
+            reporter = self.state.document.reporter
+            return [reporter.warning(msg, line=self.lineno)]
+
     def node2image(self, node, diagram):
         return node
 
@@ -44,7 +56,7 @@ def get_image_filename(self, code, format, options, prefix='seqdiag'):
     """
     Get path of output file.
     """
-    if format not in ('PNG', 'PDF'):
+    if format not in ('PNG', 'PDF', 'SVG'):
         raise SeqdiagError('seqdiag error:\nunknown format: %s\n' % format)
 
     if format == 'PDF':
@@ -119,17 +131,48 @@ def create_seqdiag(self, code, format, filename, options, prefix='seqdiag'):
         screen = builder.ScreenNodeBuilder.build(tree)
 
         antialias = self.builder.config.seqdiag_antialias
-        draw = DiagramDraw.DiagramDraw(format, screen, filename, fontmap=fontmap,
-                                       antialias=antialias)
+        draw = DiagramDraw.DiagramDraw(format, screen, filename,
+                                       fontmap=fontmap, antialias=antialias)
     except Exception, e:
         raise SeqdiagError('seqdiag error:\n%s\n' % e)
 
     return draw
 
 
+def make_svgtag(self, image, relfn, trelfn, outfn,
+                alt, thumb_size, image_size):
+    svgtag_format = """<svg xmlns="http://www.w3.org/2000/svg"
+    xmlns:xlink="http://www.w3.org/1999/xlink"
+    alt="%s" width="%s" height="%s">%s
+    </svg>"""
+
+    code = open(outfn, 'r').read().decode('utf-8')
+
+    return (svgtag_format %
+            (alt, image_size[0], image_size[1], code))
+
+
+def make_imgtag(self, image, relfn, trelfn, outfn,
+                alt, thumb_size, image_size):
+    result = ""
+    imgtag_format = '<img src="%s" alt="%s" width="%s" height="%s" />\n'
+
+    if trelfn:
+        result += ('<a href="%s">' % relfn)
+        result += (imgtag_format %
+                   (trelfn, alt, thumb_size[0], thumb_size[1]))
+        result += ('</a>')
+    else:
+        result += (imgtag_format %
+                   (relfn, alt, image_size[0], image_size[1]))
+
+    return result
+
+
 def render_dot_html(self, node, code, options, prefix='seqdiag',
                     imgcls=None, alt=None):
-    has_thumbnail = False
+    trelfn = None
+    thumb_size = None
     try:
         format = self.builder.config.seqdiag_html_image_format
         relfn, outfn = get_image_filename(self, code, format, options, prefix)
@@ -142,7 +185,6 @@ def render_dot_html(self, node, code, options, prefix='seqdiag',
         # generate thumbnails
         image_size = image.pagesize()
         if 'maxwidth' in options and options['maxwidth'] < image_size[0]:
-            has_thumbnail = True
             thumb_prefix = prefix + '_thumb'
             trelfn, toutfn = get_image_filename(self, code, format,
                                                 options, thumb_prefix)
@@ -170,15 +212,13 @@ def render_dot_html(self, node, code, options, prefix='seqdiag',
         if alt is None:
             alt = node.get('alt', self.encode(code).strip())
 
-        imgtag_format = '<img src="%s" alt="%s" width="%s" height="%s" />\n'
-        if has_thumbnail:
-            self.body.append('<a href="%s">' % relfn)
-            self.body.append(imgtag_format %
-                             (trelfn, alt, thumb_size[0], thumb_size[1]))
-            self.body.append('</a>')
+        if format == 'SVG':
+            tagfunc = make_svgtag
         else:
-            self.body.append(imgtag_format %
-                             (relfn, alt, image_size[0], image_size[1]))
+            tagfunc = make_imgtag
+
+        self.body.append(tagfunc(self, image, relfn, trelfn, outfn, alt,
+                                 thumb_size, image_size))
 
     self.body.append('</p>\n')
     raise nodes.SkipNode
@@ -215,7 +255,7 @@ def on_doctree_resolved(self, doctree, docname):
     if self.builder.name in ('gettext', 'singlehtml', 'html', 'latex'):
         return
 
-    for node in doctree.traverse(seqdiag):  
+    for node in doctree.traverse(seqdiag):
         code = node['code']
         prefix = 'seqdiag'
         format = 'PNG'